pyproject.toml
[tool.poetry]
name = "discordpy-bot-template-with-translation-functionality"
version = "0.1.0"
description = ""
authors = ["Mitchell Sharma <https://github.com/SharmaMitchell/>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
